hermes-cli/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes"
dynamic = ["version"]
description = "ML Serve cli to create base repository"
readme = "README.md"
requires-python = ">=3.6"
dependencies = [
    "annotated-types>=0.7.0",
    "anyio>=4.8.0",
    "arrow>=1.3.0",
    "binaryornot>=0.4.4",
    "pyarrow>=14.0.0",
    "certifi>=2024.12.0",
    "chardet>=5.2.0",
    "charset-normalizer>=3.4.0",
    "click>=8.1.0",
    "cookiecutter>=2.6.0",
    "fastapi>=0.115.0",
    "idna>=3.10",
    "Jinja2>=3.1.0",
    "markdown-it-py>=3.0.0",
    "MarkupSafe>=3.0.0",
    "mdurl>=0.1.2",
    "pydantic>=2.10.0",
    "pydantic_core>=2.27.0",
    "Pygments>=2.19.0",
    "python-dateutil>=2.9.0",
    "python-slugify>=8.0.0",
    "PyYAML>=6.0.0",
    "requests>=2.32.0",
    "rich>=13.9.0",
    "six>=1.17.0",
    "sniffio>=1.3.0",
    "starlette>=0.41.0",
    "text-unidecode>=1.3",
    "typer>=0.9.0",
    "typeguard>=4.4.0",
    "types-python-dateutil>=2.9.0",
    "typing_extensions>=4.12.0",
    "urllib3>=1.26.0,<2.0.0",
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "pydantic>=1.8.0",
    "python-multipart>=0.0.5",
    "loguru>=0.7.0",
    "aiohttp>=3.8.0",
    "dependency-injector>=4.45.0",
    "ddtrace==1.15.1",
    "questionary>=1.10.0",
    "black>=24.2.0",
    "pylint>=3.1.0"
]

[project.scripts]
hermes = "hermes.src.main:main"

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-html>=4.1.0"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --asyncio-mode=auto"
testpaths = [
    "tests",
]
python_files = ["test_*.py", "*_test.py"]
markers = [
    "asyncio: mark test as async",
    "api: mark test as API integration test"
]

[tool.pylint.MASTER]
ignore-paths = ["^src/template/.*$",
        "^build/.*$",
         "^.venv/.*$"
]